+++ src/services/mod.rs
@@ -2,7 +2,9 @@ pub mod about;
 pub mod balances;
 pub mod hooks;
 pub mod transactions_details;
+pub mod transactions_history;
 pub mod transactions_list;
+pub mod transactions_queued;
 pub mod tx_confirmation;
 
 #[cfg(test)]
